Fixed items outside of projects not showing up
Any items that were added to the solution directly (or in a folder that was in the solution) were not showing up in the list because we were tossing out every result that wasn't an actual Project. Now we go ahead and add the non-Project files, but ignore any invalid entries we find.