Switched to NuGet VSSDK
The built in VisualStudio DLL references were causing all sorts of trouble when trying to drop down to lower VS versions than the one I'm mainly testing on (2015 with 2013 and 2012 both installed), but using NuGet's VSSDK packages allow better targeting of older versions of libraries. This fixes the extension not working very well with VS2012.