Fixed opening non-project files
When files that weren't part of a project (added directly to a solution or open as a raw file outside the solution) were selected in the file list, a Not Implemented exception window was thrown up and VS would crash if the file was selected a second time. This catches the exception and attempts to open the file by path instead of by file reference.

Fixes #6