Trying out suggested change from VS Marketplace
This adds another layer of filtering such that if an item isn't matched by looking for each space-delimited word in the search box, we then try to see if each letter in the search term appears, in order, in the uppercase letters from item.

So, for example, if we had 3 files in our solution:
AssemblyInfo.cs
Main.cs
README

Searching for "ai" would now find the first two items since "AssemblyInfo.cs" contains "ai" in that order as uppercase letters and "Main.cs" contains "ai" per the previous rules.

I feel like this might be more intuitive if the rules were tweaked to disregard the capital letter rule, but I can see the merit in this method, so we'll try it out and see how it goes.