Drop support for VS2015 and earlier
Microsoft recommends using the meta-package for the VS SDK, but it only
supports 2017 and later. I'd like to continue to retain older version
compatibility, but it doesn't look like it's going to be an option if I
also want to use the same code/project for VS 2022 support, and it feels
more important to support newer IDEs than to support older ones.